requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # For vision transformer models

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
# opencv-python>=4.8.0  # Replaced with PIL/torchvision for better compatibility
# albumentations>=1.3.0  # Replaced with torchvision transforms
scikit-learn>=1.3.0

# Google Gemini API
google-generativeai>=0.3.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0  # Optional: for experiment tracking

# Model-specific dependencies
transformers>=4.30.0  # For some transformer architectures
einops>=0.6.1  # For tensor operations

# Evaluation metrics
scikit-image>=0.21.0
scipy>=1.11.0
